Extract property table property validation into class
